BENCH-375: Refactoring and resizing
