+++ src/components/CheckoutDetailsForm/CheckoutDetailsForm.tsx
@@ -3,7 +3,7 @@ import { RouteConstants } from 'utilities/route-constants';
 import { useLocalOrder } from 'context/OrderContext';
 import { GBPFormat } from 'utilities/GBPFormat';
 import { OrderDetails } from 'components/OrderDetails/OrderDetails';
-import { Button } from 'components/Button/Button';
+import { Button } from 'components/Buttons/Button';
 import { useNavigate } from 'react-router-dom';
 import CheckoutBurgerImg from '../../assets/icon_checkout_no_items.png';
 import { ComponentTransition } from 'components/Transitions/ComponentTransition';
