+++ src/components/OrderLoadForm/OrderLoadForm.tsx
@@ -1,5 +1,5 @@
 import './OrderLoadForm.scss';
-import { Button } from 'components/Button/Button';
+import { Button } from 'components/Buttons/Button';
 import { Error } from 'components/Error/Error';
 import { LoaderOverlay } from 'components/LoaderOverlay/LoaderOverlay';
 import { useOrder } from 'hooks/useOrder';
