+++ src/components/ProductCreateForm/ProductCreateForm.tsx
@@ -1,5 +1,5 @@
 import './ProductCreateForm.scss';
-import { Button } from 'components/Button/Button';
+import { Button } from 'components/Buttons/Button';
 import { Error } from 'components/Error/Error';
 import { LoaderOverlay } from 'components/LoaderOverlay/LoaderOverlay';
 import { useProducts } from 'hooks/useProducts';
